# server/internal/api/swagger/ui/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "File4Base API Engine",
    "description": "Enterprise-grade REST and Dynamic Relational DDL Database Engine for File4Base. Supports runtime schema migrations, table occurrences, relationships graph, dynamic Find Mode formula parsing, security ACL, and solution archiving. Every endpoint except the health probes, sign-in and the database selector requires a session token: sign in with POST /api/v1/auth/login and send the returned token as 'Authorization: Bearer <token>'. A session is bound to the database it signed in to.",
    "version": "0.5.0",
    "contact": {
      "name": "File4Base Core Team",
      "url": "https://github.com/file4base/file4base-app"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local API Service (Direct)"
    },
    {
      "url": "http://localhost:3000",
      "description": "WebDirect Nginx Reverse Proxy"
    },
    {
      "url": "/",
      "description": "Current Host"
    }
  ],
...
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Security & Users"
        ],
        "summary": "Sign in to a database and open a session",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "description": "Database not found"
          }
        },
        "security": []
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "Security & Users"
        ],
        "summary": "Sign out (revokes the session token)",
        "responses": {
          "204": {
            "description": "Session closed"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/v1/auth/session": {
      "get": {
        "tags": [
          "Security & Users"
        ],
        "summary": "Current session (user, role, database, permissions)",
        "responses": {
          "200": {
            "description": "Session details"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
    "/api/v1/security/users": {
      "get": {
        "tags": [
          "Security & Users"
        ],
        "summary": "List Users",
        "responses": {
          "200": {
            "description": "List of system users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      },
      "post": {
        "tags": [
          "Security & Users"
        ],
        "summary": "Create User",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "LoginRequest": {
        "type": "object",
        "required": [
          "database",
          "username",
          "password"
        ],
        "properties": {
          "database": {
            "type": "string",
            "example": "invoices_db"
          },
          "username": {
            "type": "string",
            "example": "alice"
          },
          "password": {
            "type": "string",
            "format": "password"
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "ok"
          },
          "database": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "token": {
            "type": "string",
            "description": "Session bearer token"
          },
          "token_type": {
            "type": "string",
            "example": "Bearer"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "owner",
              "admin",
              "user",
              "guest"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateUserRequest": {
        "type": "object",
        "required": [
          "username",
          "password",
          "role"
        ],
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "default": "user"
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "description": "RFC 9457 Problem Details for HTTP APIs",
        "properties": {
          "type": {
            "type": "string",
            "example": "about:blank"
          },
          "title": {
            "type": "string",
            "example": "Bad Request"
          },
          "status": {
            "type": "integer",
            "example": 400
          },
          "detail": {
            "type": "string",
            "example": "Invalid column format"
          },
          "instance": {
            "type": "string",
            "example": "/api/v1/schemas/tables"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Opaque session token returned by POST /api/v1/auth/login. Bound to one user and one database; expires after a period of inactivity (SESSION_TTL, default 12h)."
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing, invalid or expired session token",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            }
          }
        }
      },
      "Forbidden": {
        "description": "The session's role or layout permissions do not allow this operation",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}